rf: split into fit/apply bold workflows
